pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-proof-tutor"
version = "0.0.1"
dependencies = ["openai", "PyYAML"]
